+++ libraries/packets/src/main/java/de/oliver/fancysitula/factories/PacketFactory.java
@@ -34,18 +34,6 @@ public class PacketFactory {
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundPlayerInfoUpdatePacketImpl(actions, entries);
             }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundPlayerInfoUpdatePacketImpl(actions, entries);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundPlayerInfoUpdatePacketImpl(actions, entries);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundPlayerInfoUpdatePacketImpl(actions, entries);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundPlayerInfoUpdatePacketImpl(actions, entries);
-            }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
         }
@@ -82,18 +70,6 @@ public class PacketFactory {
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundAddEntityPacketImpl(entityId, entityUUID, entityType, x, y, z, yaw, pitch, headYaw, velocityX, velocityY, velocityZ, data);
             }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundAddEntityPacketImpl(entityId, entityUUID, entityType, x, y, z, yaw, pitch, headYaw, velocityX, velocityY, velocityZ, data);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundAddEntityPacketImpl(entityId, entityUUID, entityType, x, y, z, yaw, pitch, headYaw, velocityX, velocityY, velocityZ, data);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundAddEntityPacketImpl(entityId, entityUUID, entityType, x, y, z, yaw, pitch, headYaw, velocityX, velocityY, velocityZ, data);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundAddEntityPacketImpl(entityId, entityUUID, entityType, x, y, z, yaw, pitch, headYaw, velocityX, velocityY, velocityZ, data);
-            }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
         }
@@ -117,18 +93,6 @@ public class PacketFactory {
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundPlayerInfoRemovePacketImpl(uuids);
             }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundPlayerInfoRemovePacketImpl(uuids);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundPlayerInfoRemovePacketImpl(uuids);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundPlayerInfoRemovePacketImpl(uuids);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundPlayerInfoRemovePacketImpl(uuids);
-            }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
         }
@@ -151,18 +115,6 @@ public class PacketFactory {
             }
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundRemoveEntitiesPacketImpl(entityIds);
-            }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundRemoveEntitiesPacketImpl(entityIds);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundRemoveEntitiesPacketImpl(entityIds);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundRemoveEntitiesPacketImpl(entityIds);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundRemoveEntitiesPacketImpl(entityIds);
             }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
@@ -199,18 +151,6 @@ public class PacketFactory {
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundTeleportEntityPacketImpl(entityId, x, y, z, yaw, pitch, onGround);
             }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundTeleportEntityPacketImpl(entityId, x, y, z, yaw, pitch, onGround);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundTeleportEntityPacketImpl(entityId, x, y, z, yaw, pitch, onGround);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundTeleportEntityPacketImpl(entityId, x, y, z, yaw, pitch, onGround);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundTeleportEntityPacketImpl(entityId, x, y, z, yaw, pitch, onGround);
-            }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
         }
@@ -235,18 +175,6 @@ public class PacketFactory {
             }
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundRotateHeadPacketImpl(entityId, headYaw);
-            }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundRotateHeadPacketImpl(entityId, headYaw);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundRotateHeadPacketImpl(entityId, headYaw);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundRotateHeadPacketImpl(entityId, headYaw);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundRotateHeadPacketImpl(entityId, headYaw);
             }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
@@ -273,18 +201,6 @@ public class PacketFactory {
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundSetEntityDataPacketImpl(entityId, entityData);
             }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundSetEntityDataPacketImpl(entityId, entityData);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundSetEntityDataPacketImpl(entityId, entityData);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundSetEntityDataPacketImpl(entityId, entityData);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundSetEntityDataPacketImpl(entityId, entityData);
-            }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
         }
@@ -310,18 +226,6 @@ public class PacketFactory {
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundSetEquipmentPacketImpl(entityId, equipment);
             }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundSetEquipmentPacketImpl(entityId, equipment);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundSetEquipmentPacketImpl(entityId, equipment);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundSetEquipmentPacketImpl(entityId, equipment);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundSetEquipmentPacketImpl(entityId, equipment);
-            }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
         }
@@ -346,18 +250,6 @@ public class PacketFactory {
             }
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundSetPassengersPacketImpl(entityId, passengers);
-            }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundSetPassengersPacketImpl(entityId, passengers);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundSetPassengersPacketImpl(entityId, passengers);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundSetPassengersPacketImpl(entityId, passengers);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundSetPassengersPacketImpl(entityId, passengers);
             }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
@@ -387,18 +279,6 @@ public class PacketFactory {
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, createTeam);
             }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, createTeam);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, createTeam);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, createTeam);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, createTeam);
-            }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
         }
@@ -425,18 +305,6 @@ public class PacketFactory {
             }
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, removeTeam);
-            }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, removeTeam);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, removeTeam);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, removeTeam);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, removeTeam);
             }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
@@ -465,18 +333,6 @@ public class PacketFactory {
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, updateTeam);
             }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, updateTeam);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, updateTeam);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, updateTeam);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, updateTeam);
-            }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
         }
@@ -503,18 +359,6 @@ public class PacketFactory {
             }
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, addEntity);
-            }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, addEntity);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, addEntity);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, addEntity);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, addEntity);
             }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
@@ -543,18 +387,6 @@ public class PacketFactory {
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, removeEntity);
             }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, removeEntity);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, removeEntity);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, removeEntity);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundCreateOrUpdateTeamPacketImpl(teamName, removeEntity);
-            }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
         }
@@ -579,15 +411,6 @@ public class PacketFactory {
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundShowDialogPacketImpl(dialog);
             }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundShowDialogPacketImpl(dialog);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundShowDialogPacketImpl(dialog);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundShowDialogPacketImpl(dialog);
-            }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
         }
@@ -608,15 +431,6 @@ public class PacketFactory {
             }
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundClearDialogPacketImpl();
-            }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundClearDialogPacketImpl();
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundClearDialogPacketImpl();
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundClearDialogPacketImpl();
             }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
@@ -643,18 +457,6 @@ public class PacketFactory {
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundUpdateAttributesPacketImpl(entityId, attributes);
             }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundUpdateAttributesPacketImpl(entityId, attributes);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundUpdateAttributesPacketImpl(entityId, attributes);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundUpdateAttributesPacketImpl(entityId, attributes);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundUpdateAttributesPacketImpl(entityId, attributes);
-            }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
         }
@@ -680,18 +482,6 @@ public class PacketFactory {
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundAnimatePacketImpl(entityId, animationId);
             }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundAnimatePacketImpl(entityId, animationId);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundAnimatePacketImpl(entityId, animationId);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundAnimatePacketImpl(entityId, animationId);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundAnimatePacketImpl(entityId, animationId);
-            }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());
         }
@@ -715,18 +505,6 @@ public class PacketFactory {
             }
             case v26_1_2 -> {
                 return new de.oliver.fancysitula.versions.v26_1.packets.ClientboundBundlePacketImpl(packets);
-            }
-            case v1_21_11 -> {
-                return new de.oliver.fancysitula.versions.v1_21_11.packets.ClientboundBundlePacketImpl(packets);
-            }
-            case v1_21_9, v1_21_10 -> {
-                return new de.oliver.fancysitula.versions.v1_21_9.packets.ClientboundBundlePacketImpl(packets);
-            }
-            case v1_21_6, v1_21_7, v1_21_8 -> {
-                return new de.oliver.fancysitula.versions.v1_21_6.packets.ClientboundBundlePacketImpl(packets);
-            }
-            case v1_21_5 -> {
-                return new de.oliver.fancysitula.versions.v1_21_5.packets.ClientboundBundlePacketImpl(packets);
             }
             default ->
                     throw new IllegalArgumentException("Unsupported server version: " + ServerVersion.getCurrentVersion());